chore: cleanup redundant build commands

+++ src/micrograd/micrograd.cpp
@@ -1,9 +1,8 @@
 #include "micrograd.h"
+#include <iostream>
 
-// Constructor implementation
-Value::Value(int d) : data(d) {}
+Value::Value(double d) : data(d) {}
 
-// Operator overload implementation
 std::ostream& operator<<(std::ostream& os, const Value& value) {
     os << "Value(data: " << value.data << ")";
     return os;
